refactor: components 정리
