+++ src/components/notice/anouncement/Anouncement.style.ts
