+++ src/components/notice/info/NoticeInfo.style.ts
